feat: add all the sample use case to single module with drawer menu
