--- demo-hello-world/src/androidTest/java/com/cryptoeconomicslab/plasma_android_sdk/ExampleInstrumentedTest.kt
@@ -1,24 +1,0 @@
-package com.cryptoeconomicslab.plasma_android_sdk
-
-import androidx.test.InstrumentationRegistry
-import androidx.test.runner.AndroidJUnit4
-
-import org.junit.Test
-import org.junit.runner.RunWith
-
-import org.junit.Assert.*
-
-/**
- * Instrumented test, which will execute on an Android device.
- *
- * See [testing documentation](http://d.android.com/tools/testing).
- */
-@RunWith(AndroidJUnit4::class)
-class ExampleInstrumentedTest {
-    @Test
-    fun useAppContext() {
-        // Context of the app under test.
-        val appContext = InstrumentationRegistry.getTargetContext()
-        assertEquals("com.cryptoeconomicslab.plasma_android_sdk", appContext.packageName)
-    }
-}
