--- demo-hello-world/src/main/java/com/cryptoeconomicslab/plasma_android_sdk/App.kt
@@ -1,14 +1,0 @@
-package com.cryptoeconomicslab.plasma_android_sdk
-
-import android.app.Application
-
-/**
- * Custom Application Class
- */
-class App : Application() {
-    override fun onCreate() {
-        super.onCreate()
-
-        PlasmaClient.initialize(applicationContext)
-    }
-}
