--- demo-hello-world/src/main/res/values/strings.xml
@@ -1,3 +1,0 @@
-<resources>
-    <string name="app_name">demo-hello-world</string>
-</resources>
