+++ show_case/proguard-rules.pro
