+++ show_case/src/main/java/com/cryptoeconomicslab/show_case/App.kt
@@ -1,0 +1,15 @@
+package com.cryptoeconomicslab.show_case
+
+import android.app.Application
+import com.cryptoeconomicslab.plasma_android_sdk.PlasmaClient
+
+/**
+ * Custom Application Class
+ */
+class App : Application() {
+    override fun onCreate() {
+        super.onCreate()
+
+        PlasmaClient.initialize(applicationContext)
+    }
+}
