+++ show_case/src/main/res/layout/fragment_db.xml
