+++ show_case/src/main/res/menu/drawer_menu.xml
@@ -1,0 +1,9 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+            android:id="@+id/nav_hello_world"
+            android:title="@string/menu_hello_world"/>
+    <item
+            android:id="@+id/nav_db_sample"
+            android:title="@string/menu_db_sample"/>
+</menu>
