+++ show_case/src/main/res/values/colors.xml
