+++ show_case/src/main/res/values/styles.xml
